Removed single PR axioms referring to an obsolete GO:0043234
In order for our QC to pass community wide, we need to remove references to obsolete entities in our ontologies. Ususally, we would replace occurrences with their recommended replacements; in this case how ever, the offending axiom does not all belong into cell ontology (it should be maintained by PR!).