Classify taxon-specific follicle cells under taxon-neutral term.
Now that the distinction between the insect `follicle cell of egg
chamber` and the non-insect `follicular cell of ovary` is clear, we need
a taxon-neutral term to group both.

Initial suggestion by @dosumis was to create a new term for this
purpose, but we already have `follicular epithelial cell` (CL:0000500).
It has no definition and almost no usage except as a superclass to
`eggshell secreting cell` (CL:0000659). We thus give that term a basic
definition that avoids any taxon-specific details, and make it a
superclass of both `follicle cell of egg chamber` and `follicular cell
of ovary`.