added neocortex location to 3 basket cell types
Fixes https://github.com/obophenotype/cell-ontology/issues/1128
From reading, I think that these basket cells are in neocortex. There are basket cells in cerebellum too (and they already have a class in CL) but they seem to be another type that are multipolar. I didn't create a new class for them as they would be the only child on under cerebellum basket cells, so there isnt a need for it.