Add `gamete-nursing cell` and subclasses.
Add a new term `gamete-nursing cell` to refer to any cell that provides
cytoplasmic material to a developing gamete. Since such cells can have,
depending on the species, either a somatic or a germinal origin, this
term makes no assumption about the lineage. Instead, we also add two
subclasses to specifically refer to the germline-derived and somatic
nursing cells.

closes #1611