Add taxon restriction to `epidermal cell (sensu Arthropoda)`.
The term `epidermal cell (sensu Arthropoda)` is intended to refer to
arthropod cells only, so this PR makes that explicit with the
appropriate taxon restriction.

The label is also changed from `... (sensu arthropoda)` to `... (sensu
Arthropoda)`, for consistency with other terms with similar "sensu"
qualifications (taxons in "sensu"-type expressions are always
capitalized).

closes #342