Obsolete 'primary pigment cell'.
The term _primary pigment cell_, which refers to one type of visual
pigment cells in protostomes, is deemed to be unnecessary in CL, and is
therefore obsoleted. Suggestion is to use either the closest parental
term (_visual pigment cell (sensu Nematoda and Protostomia)_) to refer
to any kind of visual pigment cell regardless of their exact type, or a
more precise term from a taxon-specific ontology such as FBbt.

Likewise for _secondary_ and _tertiary pigment cell_.

closes #1930