Fix 'histoblast'.
Add a definition to 'histoblast'. Re-classify as a 'progenitor cell'
rather than as a 'simple columnar epithelial cell' (there is no
indication that those cells are columnar epithelial cells; there are
located within the larval epidermis, which is a type of epithelium, but
may not be epithelial cells themselves). Also add a taxon restriction to
insects.