Make 'primary neuron' teleost-specific.
'Primary neuron' (and related terms, including 'secondary neuron') was
supposed to be obsoleted for lack of a consensus on its intended
meaning, but the term may be in used to represent fish cells (following
the merging of the TAO into CL). So instead, we keep the terms but
restrict them to teleosts.