Define 'follicle stem cell' and restrict it to arthropods.
Follicle cells in insects (and _probably_ more largely in arthropods,
though evidence about that is hard to find) all derive from clearly
identified "follicle stem cells", but that does not seem to be the case
for the follicular epithelial cells found e.g. in vertebrates.

So we restrict the term 'follicle stem cell' to arthropods.