Re-classify 'R7 photoreceptor cell' as part of an ommatidium. (#2347)
The 'R7 photoreceptor cell' is explicitly classified as an 'eye
photoreceptor cell'. This is imprecise as 'eye photoreceptor cell'
refers to photoreceptors in any type of eye, while R7 photoreceptors are
found in compound eyes.

This commit classifies 'R7 photoreceptor cell' as being part of some
'ommatidium' (as the other R1, R2, ... R8 photoreceptor cells). This
automatically classifies it as a 'compount eye photoreceptor cell', as
it should be.