Building the ZFA mapping set must depend on 'all_robot_plugins'.
The rule that extracts the cross-references from the local mirror of ZFA
to build the ZFA-to-CL mapping set is using the SSSOM plugin, so we must
make sure the plugin is ready to use (in `$(TMPDIR)/plugins`) before we
reach that rule. The normal way to do that is to depend on the
'all_robot_plugins' target, which takes care of installing all plugins.

This was missed when preparing #2361 because running the entire pipeline
causes the plugins to be installed anyway before we reach the
zfa.sssom.tsv rule. But this may not be the case when trying to build
individual products.