Update ODK version used in workflows.
The `diff` workflow is still using version 1.3.0 of the ODK, but
building CL now requires version 1.5 since the preprocessing step is
making use of a ROBOT plugin and support for ROBOT plugins was
introduced in 1.5. Also, the use of plugins require *odkfull*, not
*odklite*.

Similarly, the `ontobot` workflow is still using version 1.4.3 of the
ODK, and is manually downloading the KGCL plugin. This works, but the
ODK 1.5 already includes the KGCL plugin, so upgrading the workflow to
use the latest ODK makes things a bit easier as we can just use the
plugin as provided by the ODK.

closes #2367