add report on diff between release and previous release
The report is generated by the OAK diff command using the base releases
for better comparison. The report shows new terms, new relationships,
obsolete terms, changes on synonyms and definitions. The report is
appended to the table with the ontology summary content and saved in the
reports/summary_release.md file to be used as a release note.