use cl-base.obo to generate robot release base diff
We need to download the cl-base.obo to generate the output for the OAK
diff command, so we can use the same artefact to generate the robot diff
instead of downloading another artefact. This also adds the two dependencies
for the `release-base-diff` target to make sure the files are updated.