improve the documentation about CL release workflow
Update the link to the documentation about how to update the imports
because the previous one was linking to an non-existing page.
Change to inline code syntax instead of code block the GitHub release
link because it was breaking the list numbers, making it to reset the
numbering.
Finally undo the change on the number of the last three items on the list
as mistakenly done on the previous commit.