Update ontology summary (#2329)
* enable release diff in odk config

* run release simple diff and post diff in PR

* fix release artefact uri and diff filename

* add edges, synonyms, xrefs and cl_terms reports to odk config

* update custom reports to only cl

* wip ontology content report

* add cxg and hra numbers

* revert default report sparql queries

* create custom reports for CL only

* add custom CL reports to odk config

* script to generate content summary

* add command to generate content summary as part of release

* revert changes in edges, synonyms and xrefs reports

* add instructions to add summary table in release notes

* use cl-base to run diff

* change file name and actions version

* update to commit cl-base-diff and not full release diff

* add report on diff between release and previous release

The report is generated by the OAK diff command using the base releases
for better comparison. The report shows new terms, new relationships,
obsolete terms, changes on synonyms and definitions. The report is
appended to the table with the ontology summary content and saved in the
reports/summary_release.md file to be used as a release note.

* update cl-release docs to reflect the new release process

The release notes need to be updated and this commit explains how to
fix a current issue in the OAK diff command when generating the report.

* rewording explanation in readme

Co-authored-by: Aleix Puig <94959119+aleixpuigb@users.noreply.github.com>

* add missing dependency in prepare_content_summary goal

The file is used in the rule, but it wasn't defined as a dependency, which could be used as an updated file.

Co-authored-by: Nico Matentzoglu <nicolas.matentzoglu@gmail.com>

* update the sparql queries for the custom reports

Filter out the obsolete classes and the obsolete CP namespace from the
queries not to count them in the custom reports and so to the ontology
content summary report generated for the releases.

* use cl-base.obo to generate robot release base diff

We need to download the cl-base.obo to generate the output for the OAK
diff command, so we can use the same artefact to generate the robot diff
instead of downloading another artefact. This also adds the two dependencies
for the `release-base-diff` target to make sure the files are updated.

* improve the documentation about CL release workflow

Update the link to the documentation about how to update the imports
because the previous one was linking to an non-existing page.
Change to inline code syntax instead of code block the GitHub release
link because it was breaking the list numbers, making it to reset the
numbering.
Finally undo the change on the number of the last three items on the list
as mistakenly done on the previous commit.

---------

Co-authored-by: Anita Caron <anitacaron@users.noreply.github.com>
Co-authored-by: Aleix Puig <94959119+aleixpuigb@users.noreply.github.com>
Co-authored-by: Nico Matentzoglu <nicolas.matentzoglu@gmail.com>