Make CL:2000062 part_of some 'placenta'.
The term 'placental villus capillary endothelial cell' is textually
defined as "any capillary endothelial cell that is part of a placenta",
but its logical definition restricts it more specifically to the
'placental labyrinth villous'.

But 'placental labyrinth villous', in Uberon, is said not to exist in
human (the term is intended to refer to labyrinthine-type placentas,
which exist e.g. in rodents but not in human). In effect, this means
that CL:2000062 itself, by its logical definition, cannot exist in
human.

This was likely not intended (as indicated by the fact CL:2000062 is in
fact also explicitly marked as being found in humans).

So here, we amend CL:2000062's logical definition to align it more
closely to its textual definition, by making it part_of some 'placenta'.

closes #2490