Add check for violations of taxon constraints.
We add a new test to the test suite to check that taxon constraints are
not violated.

The check is done by: (1) merging the edit file with an ontology
containing disjointness axioms over sibling taxa
(obo:ncbitaxon/subsets/taxslim-disjoint-over-in-taxon.owl), (2)
expanding any macro in the merged ontology (needed for taxon constraints
depending on RO:0002161 and RO:0002175), and (3) reasoning over the
result.

This necessitates two updates to CL's ODK configuration: adding a mirror
for the taxslim-disjoint-over-in-taxon.owl module, and making sure that
the OMO:0002000 annotation property (which defines the expansion of
RO:0002161 and RO:0002175)  is not stripped away by the imports
pipeline.