Fix TC violation imported from Uberon.
Because the TC check in Uberon had been broken between March and August
2024, a TC violation has crept in that can now be found in the Uberon
import module. That violation has already been fixed upstream in Uberon,
so it will automatically disappear the next time we refresh the imports
after the next Uberon release.

In the meantime, we hack the merged_imports module to forcefully remove
the offending axiom.