Remove obsolete OORT-related stuff from the custom Makefile.
The custom Makefile contains an entire section of rules dedicated to the
use of OORT to produce the CL release products. Those rules are
completely obsolete, as CL has not been using OORT to make its releases
for several years already.

Even if they are never used, keeping old rules and their associated
comment in the Makefile is _not_ harmless, as they needlessly make the
Makefile more difficult to read. All the comments about OORT and about a
future migration to ROBOT could easily lead someone to reasonably
conclude that CL is _still_ using OORT.

There is no reason to keep such old stuff around. If someone is
interested in CL archeology and wants to know how CL _used_ to be built
10 years ago, they can go dig through the repository's history.