Remove "non-native classes" rule.
Based on the "add back 'remove --terms non_native_classes.txt'" comment,
this rule had probably been intended as a way to allow the creation of a
kind of "-base" version of the ontology, stripped of all non-CL terms.

There are better ways to do that nowadays and that rule (which was not
used) is not needed.