Remove obsolete "templates" stuff.
All template-related rules in the custom Makefile are dependent on a
"dependencies.tsv" file that does not exist (and that has seemingly
never existed, according to the history log). The templates are supposed
to be "prepared" by a rule that downloads them from a remote location,
but the config file supposed to indicate said remote location is empty.

Clearly all this stuff is obsolete.