Allow using the ODK-generated workflow for finding pattern matches.
We add a `matches` group to the ODK configuration `patterns` section, so
that the ODK-generated Makefile will automatically a standard rule to
find matches for DOSDP patterns.

For compatibility with the existing src/patterns/data/matches directory,
the (only) `matches` item in the group is itself named `matches`. This
ensures that generated matches will be written to
src/patterns/data/matches; on the other hand, it means that the rule
generating those matches will be named `dosdp-matches-matches`, which is
quite weird.

Of note, currently the standard ODK-generated workflow cannot actually
be used, because it barfs on the ExtendedDescription pattern, which is
not supposed to be used to find matches, but the ODK does not allow to
explicitly exclude it (#2639).