Re-write custom pattern-related rules.
We cleanup the custom Makefile for all rules related to DOSDP patterns.

We remove useless stuff such as the installation of `dosdp` (already
installed in the ODK) and the production of `tmp/edit-merged.owl`, which
was used as the input for `dosdp-tools query` (not needed as dosdp-tools
can deal with imports on its own, it does not need to be fed with a
pre-merged file).

We document the rationale for the custom rules (or the _absence_ of
rationale, in the case of `gs_dosdp`) and add links to relevant tickets.

We use standard ODK variables (e.g. PATTERNDIR, ALL_PATTERN_FILES)
wherever possible.

And we regroup all the pattern-related rules in a single, clearly
labelled section.