Comment/format utility rules
Cleanup miscellaneous utility rules in the custom Makefile.

Remove the `normalise_xsd_string` rule, which is no longer useful (this
is a remnant of the transition to OWLAPI 4.5.26, if I recall correctly).

For the rules that inject IAO:0100001 annotations to the -edit file, we
rewrite them as a single rule that performs the two steps in a single
pass, for consistency with the `rm-altid` rule. The new rule is named
`add-replacedby`.

As usual, standard ODK variables are used wherever possible.