Remove the "DOSDP GH workflow".
In the three years since the "DOSDP GH workflow" has been added to the
custom Makefile, (1) the workflow has never been invoked once, and (2)
only one row has ever been added to the remote spreadsheet.

This is clearly not used and can be removed. If people want to maintain
the pattern table on the remote spreadsheet, they can still do so, and
manually copy the rows to the TSV file in this repo.

closes #2638