Break down SubClassOf axioms with an intersection as the superclass.
We have a handful of classes that are at least partially defined with
SubClassOf axioms where the superclass is an object intersection, as in
the following example:

  CL:0000104 SubClassOf: CL:0000540 and (RO:0000053 some PATO:0070026)

This is logically equivalent to individual SubClassOf axioms for each
member of the intersection:

  CL:0000104 SubClassOf: CL:0000540
  CL:0000104 SubClassOf: RO:0000053 some PATO:0070026

but the first form may not always be handled correctly and is therefore
best avoided. In particular, the OboGraphs JSON converter does not know
how to handle those axioms, and simply drops them, resulting in those
classifications being absent from the JSON artifacts.

So here, we replace known cases of SubClassOf axioms where the
superclass is an object intersection (7 found) by equivalent individual
SubClassOf axioms.

closes #2692