Compute diff on -base rather than -simple.
When computing the automated diff on pull requests, for the
"classification diff" (diff on the classified ontology), we use the
cl-base.owl product rather than the cl-simple.owl product.

Until now, we were using -simple as it was the only product that was (1)
reasoned over and (2) stripped of all "foreign" classes.

Since ODK 1.5, cl-base also fits those criteria, and it is in fact more
sensible to run the diff on that product since it is intended to be the
main product that other ontologies are supposed to be importing.

closes #1737