Remove hasAlternativeID for FMA:83808.
CL term 'cardiac muscle cell' (CL:0000746) carries a
oboInOwl:hasAlternativeId annotation to FMA:83808.

In effect, because of the way the oboInOwl:hasAlternativeId annotation
interpreted, this means that CL is declaring that FMA:83808 is
deprecated in favour of CL:0000746, and is therefore akin to an axiom
injection from CL into FMA.

Issue #1857 is asking that all oboInOwl:hasAlternativeId annotations be
converted into the modern way of representing deprecation replacement
(by a combination of owl:deprecated and IAO:0100001 annotations). But in
this instance (which is the last instance of oboInOwl:hasAlternativeId
in CL), because the (supposedly) deprecated term is in FMA, it is up to
FMA to act. On the CL side, the correct thing to do is simply to remove
the annotation, which is what we do here.

closes #1857