Fix bogus xml:base attribute in XML catalog.
The `group` element in the XML catalog contains a `xml:base` attribute
with an empty value (`xml:base=""`). It is unclear whether an empty
value is legal (though the XML Base specification [1] does not seem to
explicitly forbid it), but in any case it does not make any sense, and
some XML Catalog parsers barf upon encountering such an attribute. So
here we simply remove it.

This has (as expected) no effect on the parsing of the catalog by both
Protégé and ROBOT.

[1] https://www.w3.org/TR/xmlbase/