Layers of the GI tract
Figure depicts the layers present in the gastrointestinal tract. Presence and characteristics of these layers depends on the organ within the system. Adventitia is present on the outermost layer instead of serosa in retroperitoneal (fixed) structures. 

Lamina propria is referred to as circular muscle in this figure. Muscularis mucosae is referred to as longitudinal muscle fibres. 

Right hand figure depicts a close up of the intestinal villus also referencing the crypts of Lieberkuhn as ‘mouth of crypt’. Lacteals are the central lymphatic vessels in the villi of the small intestine and are illustrated in this figure as a green line in the cross section of a villus.