Prevent update_repo from overwriting the QC workflow.
CL has a customised QC workflow, in which we additionally check for
taxon constraint violations. Whenever the repository is updated (e.g. to
reflect changes in cl-odk.yml after adding a new component), our QC
workflow file should _not_ be overwritten by the ODK-generated one.

When upgrading to a newer major version of the ODK, if the never version
brings changes to the standard QC workflow, it will be up to a CL
"pipeline engineer" to port those changes if necessary.