Add the CLM-CL component.
We add a new component obtained from the Cell Markers Ontology
(https://github.com/Cellular-Semantics/CellMark).

For now, it is refreshed unconditionally as part of all workflows, as
for the other similar components (the HRA subset and the CellxGene
subset). Later we will need to uncouple the refresh from at least the QC
workflow.