Only refresh external resources upon MIR=true.
This commit make sure that external resources (e.g. mapping sets and
components that are maintained elsewhere, and simply used as they are in
CL) are only refreshed when the MIR Make variable is set to true.

Importantly, when running the QC workflows MIR is false, which means
that those resources will then _not_ be refreshed as part of the QC.
This is on purpose. Those external resources could at any time introduce
changes that could break CL. QC should test changes that happen entirely
within CL.

With that changes, it is up to CL's editors/maintainers to refresh the
external resources _when they want to do so_. For that, all they need to
do is to make sure that MIR is set to true. For example, this is done
automatically when calling `make refresh-imports`.

closes #2644