Include the taxon subset tags in the main release product.
Now that we can generate tag files for the taxon subsets, we merge those
into the main cl.owl release artefact.

This requires that we no longer build the taxon subsets from the same
release artefact, to avoid an obvious chicken-and-egg problem. So we now
generate the subsets from cl-full.owl instead (which is, in effect,
almost the same thing as cl.owl, modulo the ontology annotation).