Produce human- and mouse-specific subsets.
Add two new subsets: human-view and mouse-view. They are automatically
generated by applying the taxon constraints declared in the ontology to
exclude any class that can be inferred not to exist in human and mouse,
respectively.

This is the same strategy as used to produce similar subsets in Uberon,
and is reusing the `create-species-subset` custom ROBOT command
developed to that effect in the Uberon ROBOT plugin.