Remove rules for diffs that are no longer used.
We remove a bunch of rules in the custom Makefile that were intended to
generate some custom diffs. Those rules are never called from amywhere,
so the diffs are only generated when they are explicitly called by a
user on the command line -- which means they are probably never used at
all, especially now that we are producing and *publishing* more complete
diffs on every release.

closes #2641