Remove useless label annotation assertion for oboInOwl:hasDbXref.
The -edit file injects its own label for the oboInOwl:hasDbXref
property. There is no reason do to anything like that, and the label
that is injected is different from the labels we get from the imports.

Thise causes needless rewriting of the comment associated with the
annotation property, depending on whether the -edit file is written by
Protégé or by ROBOT.

Removing the annotation altogether.