Let the ODK publish the main CL mapping set.
In ODK 1.6, the `release_mappings` feature, which was broken in ODK 1.5,
has been fixed, so we no longer need to take any step to ensure the CL
mapping set is treated as a release artefacts, apart from declaring it
as such in the ODK configuration.

This means that we can remove the `deploy_release` target (automatically
called as part of the `cl` target, if `DEPLOY_GH` is enabled), because
it is now strictly equivalent to the built-in `public_release` target.