Update the Uberon ROBOT plugin.
The Uberon plugin has been moved from my personal GitHub space
(https://github.com/gouttegd/) to the "obophenotype" organisation
(https://github.com/obophenotype/). So we update the location we are
downloading the plugin from, and get the latest version.

There was also one remaining use of the now deprecated
`uberon:obo-export` command, which we replace by a call to the standard
`convert` command with the (new in ROBOT 1.9.8) `--clean-obo` option.