Structure of the cerebellar cortex
doi: 10.3390/anatomia2010005
paper legend: Structure of the cerebellar cortex. In the outermost part is the molecular layer, composed of the dendrites of Purkinje cells, the axons of granule cells (parallel fibers), and the climbing fibers; basket and stellate cells are also located in this layer. Deep below is the Purkinje layer, formed by the somas of these neurons. The inner layer of the cerebellar cortex is the granular layer, made up of granule cells and other interneurons, such as Golgi cells.
Creative Commons licence