Support temporary IDs. (#3334)
This commit ports to CL the temporary ID management system already in
use in Uberon, for allowing AI agents (or in fact any editor, AI or not)
to use “temporary” IDs that are to be later replaced by definitive IDs.

It allocates two new ID ranges:

* the `Temporary IDs` range (CL:99NNNNN): any ID picked within that
  range is treated as a temporary ID, slated for later replacement by a
  definitive ID;
* the `Automation` range (CL:0020000–CL:0120000): this is a “normal” ID
  range (IDs picked within that range are not treated differently from
  any other ID), but is reserved for automated processes – notably, for
  the process that replaces temporary IDs by definitive ones.

It also adds a new GitHub Action workflow, `allocate-definitive-ids`,
which when triggered finds all temporary IDs (all IDs in the CL:99NNNNN
range) in the cl-edit.owl file and replaces them by newly allocated IDs
from the `Automation` range. That workflow can be used in two different
ways:

(A) Manual trigger: When a PR contains temporary IDs, a CL maintainer
can manually trigger the `allocate-definitive-ids` workflow on the PR’s
underlying branch, _just before_ merging the PR into the master branch.
This will automatically add a commit to the PR, in which temporary IDs
are replaced by definitive ones. The PR should then be merged _as soon
as possible_ once that commit has been added.

(B) Automatic trigger: If a PR containing temporary IDs is merged into
the master branch (with the temporary IDs still present), the
`allocate-definitive-ids` workflow will automatically be triggered and
push a new commmit to replace the temporary IDs by definitive ones,
directly on the master branch.