Make sure all subsets are in RDF/XML format.
The production of the SemSQL version of an artefact requires that said
artefact be in RDF/XML format, because that is the only format supported
by rdftab.

The *-view subsets currently are exported in OFN format, so we need to
change that.

Of note, this will cause a massive diff at the next release, since all
the committed subsets will be completely rewritten from OFN to RDF/XML!
But maybe those subsets should not even have been committed in the first
place anyway.